+++ tket/src/Circuit/latex_drawing.cpp
@@ -37,7 +37,9 @@ void add_latex_for_command(LatexContext& context, const Command& command) {
   const Op_ptr op = command.get_op_ptr();
   switch (op->get_type()) {
     case OpType::CnRy:
-    case OpType::CnX: {
+    case OpType::CnX:
+    case OpType::CnY:
+    case OpType::CnZ: {
       int target_index = line_ids.at(args.back());
       args.pop_back();
       for (const UnitID& u : args) {
@@ -51,8 +53,14 @@ void add_latex_for_command(LatexContext& context, const Command& command) {
             << "\\gate{\\text{"
             << get_op_ptr(OpType::Ry, op->get_params())->get_name(true)
             << "}} & ";
-      } else {
+      } else if (op->get_type() == OpType::CnX) {
         lines.at(target_index).buffer << "\\targ{} & ";
+      } else if (op->get_type() == OpType::CnY) {
+        lines.at(target_index).buffer << "\\gate{\\text{"
+                                      << get_op_ptr(OpType::Y)->get_name(true)
+                                      << "}} & ";
+      } else {
+        lines.at(target_index).buffer << "\\control{} & ";
       }
       lines.at(target_index).depth++;
       break;